CTE example
This is an example of a CTE and usage of the CTE...the first part of this example shows a CTE that I created to retrieve the "To Inventory" value from an ERP (the @CUTOFFDATE value had been supplied by the parameter value selected within a SSRS report). The second part of this example shows the query that I had written, which uses the result sets from CTEs, the TOINVQTY had been joined onto "WIPVALUES" where the job numbers had matched and "ISNULL(TOINVQTY.ToInv,0) AS 'ToInv_Qty'," had been included in the query so that the value from the "TOINVQTY" CTE would be retrieved.